chore: rename files, format and fix routes
