+++ src/api/routes/admin/role/get-all-users.ts
@@ -1,8 +1,10 @@
 import { Request, Response } from "express";
-import UserService from "../../../../services/user"
-export default async (req: Request, res: Response) =>{
+
+import UserService from "../../../../services/user";
+
+export default async (req: Request, res: Response) => {
   const userService = req.scope.resolve("userService") as UserService;
   const users = await userService.list();
 
-  res.json({users});
-}
+  res.json({ users });
+};
