+++ src/api/routes/admin/role/get-permission.ts
@@ -1,10 +1,12 @@
 import { Request, Response } from "express";
+
 import PermissionService from "../../../../services/permission";
 
-export default async (req: Request, res: Response) =>{
-  const permissionService = req.scope.resolve("permissionService") as PermissionService;
+export default async (req: Request, res: Response) => {
+  const permissionService = req.scope.resolve(
+    "permissionService"
+  ) as PermissionService;
   const permission = await permissionService.listPermission();
 
-  res.json({permission});
-}
-
+  res.json({ permission });
+};
