+++ src/api/routes/admin/role/get-role-by-id.ts
@@ -1,10 +1,10 @@
 import { Request, Response } from "express";
 import RoleService from "../../../../services/role";
 
-export default async (req: Request, res: Response) =>{
-    const id = req.params.id as string;
+export default async (req: Request, res: Response) => {
+  const id = req.params.id as string;
   const roleService = req.scope.resolve("roleService") as RoleService;
-  const role = await roleService.retrieve1(id);
+  const role = await roleService.retrieve(id);
 
-  res.json({role});
-}
+  res.json({ role });
+};
