+++ src/api/routes/admin/role/get-roles.ts
@@ -1,9 +1,10 @@
 import { Request, Response } from "express";
+
 import RoleService from "../../../../services/role";
 
-export default async (req: Request, res: Response) =>{
+export default async (req: Request, res: Response) => {
   const roleService = req.scope.resolve("roleService") as RoleService;
   const role = await roleService.listRoles();
 
-  res.json({role});
-}
+  res.json({ role });
+};
